Player controller done!
The basics of the player controller is done!

There is a cooldown for dashing and for attacking.

There is one bug that I HAVE to fix however... Dashing causes the player controls to... Mess up?

Not sure what causes it, I suspect it may be the controller mode?
Not sure, will investigate :)