Updated Player Interface logic.
Updated how interactable objects work:
There is now an inherited interface derived from the 'BP_Parent_Usable' class, which has an interface.

Just make a child class of 'BP_Parent_Usable' and mark the interface function as an event. And you are good to go :)

I also made a new door class that is a child of this parent class :)

AAAND I may have accidentally deleted the player spawner... oops...

So there is now a new spawner blueprint for the player, as unreal refuses to let me name the blueprint the same ._.