Added 12Hour timer! Sorta...
Got it working... Sorta...
So... I'm Norwegian, I don't understand how boundaries between AM and PM work, like...

is it 12AM in the middle of the morning? Is 12:01AM valid? Or would it be 00:01PM... I don't... understand...

24Hour > 12Hour...

I'll figure out the boundary system later. For now, this seems to work... I think